ranWord Method
Rough draft
