+++ package/utilities/__init__.py
@@ -1,0 +1,3 @@
+from package.utilities.word import Word
+
+__all__ = ['Word']
